replacing widening pointer conversion to be able to compile with c++03
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@26055 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ topcaf/modules/WaveTimingV3Module/src/WaveTimingV3Module.cc
@@ -104,8 +104,8 @@ void WaveTimingV3Module::event()
 
     //      m_tmp_h->Smooth(2);
     int search_peaks_found = spec->Search(&m_tmp_h, m_sigma, "nodraw", 0.1);
-    double* xpos = spec->GetPositionX();
-    double* ypos = spec->GetPositionY();
+    float* xpos = spec->GetPositionX();
+    float* ypos = spec->GetPositionY();
 
     //Create TOPDigit
     int channelID = -9;
